requirements.txt
# Agent Zero v8.0 - 核心依赖
# ==========================================
# 这是运行 Agent Zero 所需的最小依赖集

# ============================================================
# LLM 框架
# ============================================================
# 基于实际安装版本（Python 3.13.9 测试通过）
langchain>=1.0.0
langgraph>=0.2.0
langchain-openai>=1.0.0
langchain-anthropic>=1.0.0
langchain-community>=0.3.0

# ============================================================
# 向量数据库
# ============================================================
chromadb>=0.4.22

# ============================================================
# 文档处理
# ============================================================
unstructured>=0.12.0
pymupdf>=1.23.0
python-magic-bin>=0.4.14; sys_platform == 'win32'
python-magic>=0.4.27; sys_platform != 'win32'
python-docx>=1.1.0

# ============================================================
# 数据验证和序列化
# ============================================================
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ============================================================
# 模板引擎
# ============================================================
jinja2>=3.1.2

# ============================================================
# UI 框架
# ============================================================
streamlit>=1.30.0
plotly>=5.0.0

# ============================================================
# 版本控制
# ============================================================
gitpython>=3.1.40

# ============================================================
# 工具库
# ============================================================
# HTTP 客户端
httpx>=0.25.0
requests>=2.31.0

# 异步支持
aiohttp>=3.9.0

# 重试机制
tenacity>=8.2.3

# 日志
loguru>=0.7.0

# 配置管理
python-dotenv>=1.0.0

# YAML 处理
pyyaml>=6.0.1

# ============================================================
# 搜索工具
# ============================================================
tavily-python>=0.3.0

# ============================================================
# 代码质量
# ============================================================
black>=25.12.0
ruff>=0.1.9
